refactor: renamed the openapi_router to just app, as the router part is a feat of the lib
