+++ src/lib/api_app/src/api/mod.rs
@@ -3,7 +3,7 @@ use lambda_http::Request;
 use log;
 
 use self::routes::get_router_response;
-use openapi_router::{
+use app::{
     responses::APIRoutingResponse,
     router::{parse_router_request, ParsedRequest},
 };
