+++ src/lib/api_app/src/api/routes/application.rs
@@ -3,7 +3,7 @@ use reqwest::Response;
 use serde_json::{json, Value as JsonValue};
 use std::{env, fs};
 
-use openapi_router::{
+use app::{
     requests::engage_many_plain_requests,
     responses::{resolve_external_service_bad_response, APIResponse, APIRoutingResponse},
     router::ParsedRequest,
