+++ src/lib/api_app/src/api/routes/jmf/mod.rs
@@ -1,8 +1,6 @@
 use std::env;
 
-use openapi_router::{
-    requests::post_json_request, responses::APIResponse, router::ParsedRequest,
-};
+use app::{requests::post_json_request, responses::APIResponse, router::ParsedRequest};
 use utils::api::get_default_headers;
 
 pub mod models;
