+++ src/lib/api_app/src/api/routes/testbed/mod.rs
@@ -7,7 +7,7 @@ use serde_json::Value as JSONValue;
 
 use crate::api::routes::application::get_external_service_bad_response;
 
-use openapi_router::{
+use app::{
     responses::{APIResponse, APIRoutingError, APIRoutingResponse},
     router::ParsedRequest,
 };
