+++ src/lib/api_app/src/api/routes/testbed/productizers/job/mod.rs
@@ -4,7 +4,7 @@ use std::{cmp::Ordering, collections::hash_map::DefaultHasher, env, hash::Hasher
 
 use super::parse_testbed_request_headers;
 
-use openapi_router::{
+use app::{
     requests::engage_many_json_requests,
     responses::{
         resolve_external_service_bad_response, APIResponse, APIRoutingError, APIRoutingResponse,
