+++ src/lib/api_app/src/api/routes/testbed/productizers/person/basic_information/mod.rs
@@ -1,4 +1,4 @@
-use openapi_router::{
+use app::{
     router::ParsedRequest,
     responses::APIResponse,
 };
