+++ src/lib/api_app/src/api/routes/testbed/productizers/person/job_applicant_profile/mod.rs
@@ -1,4 +1,4 @@
-use openapi_router::{
+use app::{
     router::ParsedRequest,
     responses::APIResponse,
 };
