+++ src/lib/api_app/src/lib/app/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod requests;
+pub mod responses;
+pub mod router;
