+++ src/lib/api_app/src/lib/app/src/router.rs
@@ -1,8 +1,34 @@
+use futures::Future;
 use http::header::HeaderMap;
 use lambda_http::{aws_lambda_events::query_map::QueryMap, Body, Request, RequestExt};
 
+use super::responses::APIResponse;
+pub use openapi_router_derive::*;
 use utils::strings;
 
+// Interface for OpenAPI operations handler
+pub trait OpenApiRouter {
+    type FutureType: Future<Output = APIResponse> + 'static;
+
+    fn get_operation(
+        &self,
+        operation_id: String,
+        parsed_request: ParsedRequest,
+    ) -> Box<dyn FnOnce() -> Self::FutureType + Send>;
+
+    fn run_operation(
+        &self,
+        operation_id: String,
+        parsed_request: ParsedRequest,
+    ) -> Self::FutureType {
+        let closure = self.get_operation(operation_id, parsed_request);
+        closure()
+    }
+}
+
+/**
+ * Request input for the router
+ */
 pub struct ParsedRequest {
     pub path: String,
     pub method: String,
