--- src/lib/api_app/src/lib/openapi_router/src/lib.rs
@@ -1,30 +1,0 @@
-use futures::Future;
-
-pub use openapi_router_derive::*;
-
-pub mod requests;
-pub mod responses;
-pub mod router;
-
-use responses::APIResponse;
-use router::ParsedRequest;
-
-// Interface for OpenAPI operations handler
-pub trait OpenApiRouter {
-    type FutureType: Future<Output = APIResponse> + 'static;
-
-    fn get_operation(
-        &self,
-        operation_id: String,
-        parsed_request: ParsedRequest,
-    ) -> Box<dyn FnOnce() -> Self::FutureType + Send>;
-
-    fn run_operation(
-        &self,
-        operation_id: String,
-        parsed_request: ParsedRequest,
-    ) -> Self::FutureType {
-        let closure = self.get_operation(operation_id, parsed_request);
-        closure()
-    }
-}
